refactor: rearranging the roles of the plugin and the extension
